test_reg_per_tok/requirements.txt
django
psycopg2-binary
python-decouple
pillow
djangorestframework
dj-rest-auth